Add TokenParser generalization and renamed file with error classes
