+++ lib/ruber_dialog.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "ruber_dialog/version"
-require_relative "ruber_dialog/parser/parser"
+require_relative "ruber_dialog/parser/errors"
 require_relative "ruber_dialog/parser/tokens"
 require_relative "ruber_dialog/parser/character_block_parser"
