+++ lib/ruber_dialog/parser/character_block_parser.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "parser"
+require_relative "errors"
 require_relative "tokens"
 require_relative "character_parser"
 module RuberDialog
