+++ lib/ruber_dialog/parser/character_parser.rb
@@ -1,48 +1,17 @@
 # frozen_string_literal: true
 
-require_relative "parser"
+require_relative "errors"
 require_relative "tokens"
+require_relative "parser"
 
 module RuberDialog
   module Parser
-    # single character parser from string
-    class CharacterParser
-      attr_accessor :forbidden_expressions, :reserved_names
-
+    # Single Character parser from string
+    class CharacterParser < TokenParser
       def initialize(forbidden_expressions: [], reserved_names: [])
-        @forbidden_expressions = forbidden_expressions
-        @reserved_names = reserved_names
-      end
-
-      def validate_forbidden_expressions(content)
-        errors = []
-
-        @forbidden_expressions.each do |expression|
-          if content&.include?(expression)
-            errors << ValidationError.new(content.index(expression), "Forbidden symbol '#{expression}'")
-          end
-        end
-        errors
-      end
-
-      def validate_reserved_names(content)
-        errors = []
-
-        @reserved_names.each do |reserved_name|
-          if content&.start_with?(reserved_name)
-            errors << ValidationError.new(0, "Use of reserved name (#{reserved_name}) as a character name is forbidden")
-          end
-        end
-        errors
-      end
-
-      private :validate_forbidden_expressions, :validate_reserved_names
-
-      def validate(content)
-        errors = validate_forbidden_expressions(content)
-        errors_reserved_names = validate_reserved_names(content)
-        errors.push(*errors_reserved_names)
-        errors.sort_by(&:position)
+        setup_forbidden_expression_error { |fe| "Forbidden symbol '#{fe}'" }
+        setup_reserved_name_error { |rn| "Use of reserved name (#{rn}) as a character name is forbidden" }
+        super(forbidden_expressions, reserved_names)
       end
 
       def parse(content)
